Add crucifix to shape effect

+++ xLights/effects/EffectManager.h
@@ -77,7 +77,7 @@ public:
         }
 
         RenderableEffect *GetEffect(const std::string &str) const;
-        int GetLastEffectId() const { return size() - 1; }
+        int GetLastEffectId() const { return (int)size() - 1; }
 
         int GetEffectIndex(const std::string &effectName) const;
         const std::string &GetEffectName(int idx) const;
